css/css-scroll-snap/scroll-snap-stop-always.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
div {
  position: absolute;
}
#scroller {
  width: 400px;
  height: 400px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
}
#space {
  left: 0px;
  top: 0px;
  width: 2100px;
  height: 2100px;
}
.target {
  width: 50px;
  height: 50px;
  scroll-snap-align: start;
}
.always-stop {
  left: 100px;
  top: 0px;
  scroll-snap-stop: always;
}
.closer {
  left: 200px;
  top: 0px;
}
</style>

<div id="scroller">
  <div id="space"></div>
  <div class="target always-stop"></div>
  <div class="target closer"></div>
</div>

<script>
test(() => {
  var scroller = document.getElementById("scroller");
  scroller.scrollBy(300, 0);
  assert_equals(scroller.scrollLeft, 100);
  assert_equals(scroller.scrollTop, 0);
}, 'A scroll with intended direction and end position should not pass a snap ' +
   'area with scroll-snap-stop: always.')
</script>
